third_party/WebKit/LayoutTests/external/wpt/mixed-content/optionally-blockable/no-opt-in/same-host-http/audio-tag/top-level/swap-scheme-redirect/no-opt-in-allows.https-expected.txt

This is a testharness.js-based test.
FAIL opt_in_method: no-opt-in
                                 origin: same-host-http
                                 source_scheme: https
                                 context_nesting: top-level
                                 redirection: swap-scheme-redirect
                                 subresource: audio-tag
                                 expectation: allowedassert_equals: The triggered event should match 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
